court-specific/atlanta-municipal/atlanta_court_scraper/pyproject.toml
[project]
name = "atlanta-court-scraper"
version = "0.1.0"
description = "Client for searching Atlanta Municipal Court cases via Benchmark system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "python-dotenv",
    "beautifulsoup4",
]

[project.scripts]
atlanta-court = "atlanta_court.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipykernel",
    "pytest",
    "pytest-cov",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["src/atlanta_court"]
